refactor(mutation-kill): extract retry/downgrade module, merge Parallelism docs
Closes #1925
Closes #1926

Extracts the retry-then-downgrade policy (is_gateway_class_error,
make_retrying_headless_call, GenerationExhausted, DowngradeEvent,
make_downgrade_audit_hook, #1908) out of mutation_kill_shared.py into
its own mutation_kill_retry.py — that module had grown into a
five-concern grab-bag where a git-revert fix and a downgrade-ladder fix
shared the same review blast radius. The new module's only dependency
on mutation_kill_shared is run_claude_headless/resolve_fallback_model,
accessed via the module object so existing monkeypatch-based tests
keep working unchanged. Test coverage for the retry policy moves to a
parallel test_mutation_kill_retry.py.

Also merges mutation-kill.md's two top-level Parallelism sections
(## Parallelism and ## Parallel execution (Phase 4)) into one, with
the sub-agent fan-out material as a ### subsection, so --concurrency's
meaning is stated once instead of twice.

Fixes surfaced by a 14-agent review panel (re-verified by a 14-agent
closing pass): stale mutation_kill_shared docstring cross-references
left behind by the extraction (3 files), a broken markdown anchor from
the section merge, a stale ADR 0035 cross-reference to the old heading
name, and a stale test-docstring pointer to the pre-split test file.
