+++ plugins/dev-team/agents/mutation-kill.md
@@ -39,7 +39,7 @@ structurally-unkillable code. Everything else is delegated.
 - `--all` — run all files in survivor-count order (highest first).
 - `--report <path>` — load an existing report instead of running the tool (first round only).
 - `--max-rounds <n>` — maximum rounds per file (default: 5).
-- `--concurrency <n>` — parallel files via git worktrees when using `--all` (default: 2; max = physical cores − 2); `--parallel <n>` — Phase 4 sub-agent fan-out via the Agent tool (in-process, no worktrees; see [Parallel execution (Phase 4)](#parallel-execution-phase-4)).
+- `--concurrency <n>` — parallel files via git worktrees when using `--all` (default: 2; max = physical cores − 2); `--parallel <n>` — Phase 4 sub-agent fan-out via the Agent tool (in-process, no worktrees; see [Sub-agent fan-out within a file](#sub-agent-fan-out-within-a-file---parallel)).
 - `--skip-static-mutants` — opt-in, default OFF; JS/TS (Stryker) path only, agent-parsed (no argparse CLI for JS/TS). See [Static-mutant skip](../skills/mutation-testing/references/languages/javascript-stryker.md#static-mutant-skip-skip-static-mutants) for the full contract.
 
 ## Deterministic mechanics are scripted — you own generation and exclusion judgment
@@ -54,7 +54,8 @@ re-describe or re-implement their mechanics:
 | `mutation_kill_loop.py` | The C#/Stryker.NET per-file loop: scoped run → score → survivor check → **your** generation → duplicate-guard → insert-before-class-close → build → test → commit-on-green / revert-on-failure → no-improvement stop. Delegates DOTNET_ROOT + `.sln` hide/restore to the wrapper. Config parsing and `run_for_file` orchestration only — insertion mechanics and headless generation live in the sibling scripts below. |
 | `mutation_kill_insert.py` | C# test-method insertion mechanics: detect-or-refuse — duplicate-name guard, and inserting generated methods before the test class's closing brace (refuses on a file-scoped namespace or non-4-space indentation rather than risk a mis-insertion). |
 | `mutation_kill_insert_python.py` | Python/pytest test-function insertion mechanics — the Python mirror of `mutation_kill_insert.py`: detect-or-refuse duplicate-name guard, and appending generated functions at the end of the file (refuses on a class-based test file rather than risk a mis-insertion). |
-| `mutation_kill_shared.py` | Cross-language loop mechanics shared verbatim between the two loops: env-var timeout parsing, `git_revert`/`git_reset_and_revert`/`git_commit`, the "no improvement across rounds" stop predicate, the `claude --print` headless-generation glue (`resolve_model`, `strip_code_fences`, `claude_cli_available`, `CLAUDE_CLI`, `run_claude_headless`), the unified `InsertOutcome`/`InsertionRefused` result types both loops' insertion scripts return, and the retry-then-downgrade policy (`is_gateway_class_error`, `make_retrying_headless_call`, `DowngradeEvent`, `GenerationExhausted`) plus its audit hook (`make_downgrade_audit_hook`). |
+| `mutation_kill_shared.py` | Cross-language loop mechanics shared verbatim between the two loops: env-var timeout parsing, `git_revert`/`git_reset_and_revert`/`git_commit`, the "no improvement across rounds" stop predicate, the `claude --print` headless-generation glue (`resolve_model`, `resolve_fallback_model`, `strip_code_fences`, `claude_cli_available`, `CLAUDE_CLI`, `run_claude_headless`), and the unified `InsertOutcome`/`InsertionRefused` result types both loops' insertion scripts return. |
+| `mutation_kill_retry.py` | The retry-then-downgrade policy on repeated headless-generation failures (`is_gateway_class_error`, `make_retrying_headless_call`, `DowngradeEvent`, `GenerationExhausted`) plus its audit hook (`make_downgrade_audit_hook`) — extracted out of `mutation_kill_shared.py` once that module grew into a five-concern grab-bag; its only dependency on `mutation_kill_shared.py` is `run_claude_headless`/`resolve_fallback_model`. |
 | `mutation_safety_gate.py` | Shared deny-list scan + refuse-on-match guard against prompt-injection payloads in generated test code, plus the commit audit-trailer (`append_generator_trailer`). |
 | `mutation_kill_headless.py` | The C#-specific headless generation + `--headless` CLI/entry point: builds the C#-flavored generation prompt and dispatches `mutation_kill_loop.run_for_file`. Imports its generic (non-C#-specific) helpers (`resolve_model`, `claude_cli_available`, `CLAUDE_CLI`, `run_claude_headless`) from `mutation_kill_shared.py` rather than defining them — `mutation_kill_loop_python.py` imports the same names directly from `mutation_kill_shared.py`, not through this module, so the Python loop carries no dependency on the C#/Stryker.NET stack. |
 | `mutation_kill_loop_python.py` | The Python/mutmut per-file loop — same contract, adapted for pytest: scoped `mutmut run` (clears stale `.mutmut-cache` first) → score via `mutation_report` junitxml support → **your** generation → duplicate-guard → append-at-end-of-file → `py_compile` → scoped `pytest` → commit-on-green / revert-on-failure → no-improvement stop. Insertion mechanics live in `mutation_kill_insert_python.py`; reuses `mutation_kill_shared.py`'s git/timeout/stop-predicate/headless-generation mechanics rather than duplicating them. |
@@ -571,7 +572,7 @@ build-artifacts directory). Concurrent runs saturate CPU/RAM fast — honor
 cores − 2). For unattended CI, `stryker_shard_pipeline.py` provides the
 compounding-worktree, forced-`--headless` alternative described above.
 
-## Parallel execution (Phase 4)
+### Sub-agent fan-out within a file (`--parallel`)
 
 `--concurrency` fans **files** out across git worktrees. `--parallel <n>` fans
 **sub-agents** out **within** a file's Phase-4 survivor set, using the Agent
