+++ tests/scripts/test_mutation_kill_headless.py
@@ -33,6 +33,7 @@ from _mutation_test_helpers import (
 sys.path.insert(0, str(SCRIPTS_DIR))
 
 import mutation_kill_headless as headless
+import mutation_kill_retry as retry
 import mutation_kill_shared as shared
 
 
@@ -310,7 +311,7 @@ def test_main_returns_exit_code_5_when_generation_exhausted_propagates(
     monkeypatch.setattr(headless, "claude_cli_available", lambda: True)
 
     def boom(*a, **k):
-        raise shared.GenerationExhausted("foo.cs (round 3) exhausted its retry budget")
+        raise retry.GenerationExhausted("foo.cs (round 3) exhausted its retry budget")
 
     monkeypatch.setattr(headless, "run_for_file", boom)
     config_path = _write_config(tmp_path)
@@ -332,7 +333,7 @@ def test_main_returns_exit_code_5_when_generation_exhausted_propagates(
 
 # =============================================================================
 # Scenario: make_headless_generator wires through
-# mutation_kill_shared.make_retrying_headless_call (#1908, Slice 3 Step 3.2)
+# mutation_kill_retry.make_retrying_headless_call (#1908, Slice 3 Step 3.2)
 # — round tracking, per-closure isolation, and GenerationExhausted
 # propagation. The retry/downgrade behavior itself is covered exhaustively
 # in test_mutation_kill_shared.py; this file only proves the wiring.
@@ -379,7 +380,7 @@ def test_make_headless_generator_propagates_generation_exhausted(
         "opus", log=lambda _: None, sleep=lambda _s: None
     )
 
-    with pytest.raises(shared.GenerationExhausted):
+    with pytest.raises(retry.GenerationExhausted):
         generate("Foo.cs", [_mutant("Survived")], "src", "tests")
 
 
@@ -448,7 +449,7 @@ def test_make_headless_generator_label_override_reflects_a_downgrade(
     monkeypatch.delenv("DEV_TEAM_MUTATION_FALLBACK_MODEL", raising=False)
     gw = lambda: RuntimeError("claude CLI failed (exit 1): 502 Bad Gateway")
     sequenced_run_claude_headless(monkeypatch, shared, gw(), gw(), gw(), gw(), "ok-on-sonnet")
-    on_downgrade, get_label_override = shared.make_downgrade_audit_hook()
+    on_downgrade, get_label_override = retry.make_downgrade_audit_hook()
     generate = headless.make_headless_generator(
         "opus", log=lambda _: None, on_downgrade=on_downgrade, sleep=lambda _s: None
     )
